requirements.txt
asgiref==3.7.2
black==24.1.1
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.3
defusedxml==0.8.0rc2
distlib==0.3.8
Django==4.2.9
django-environ==0.11.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-redis
filelock==3.13.1
flake8==7.0.0
identify==2.5.33
idna==3.6
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1
virtualenv==20.25.0
